fix(metrics_linux): 修复磁盘用量计算时的类型不匹配问题
将st.Bsize从int转换为int64，匹配diskUsageMBFromBlocks的参数类型，消除潜在的类型转换警告和计算错误

+++ internal/cfprobe/metrics_linux.go
@@ -278,7 +278,7 @@ func diskUsageLinux() (uint64, uint64) {
 		if err := syscall.Statfs(mountPoint, &st); err != nil {
 			return
 		}
-		size, used, ok := diskUsageMBFromBlocks(st.Blocks, st.Bfree, st.Bsize)
+		size, used, ok := diskUsageMBFromBlocks(st.Blocks, st.Bfree, int64(st.Bsize))
 		if !ok {
 			return
 		}
